ファイル収集と表示用に分けた
