+++ 08.ls_object/ls.rb
@@ -1,6 +1,7 @@
 #!/usr/bin/env ruby
 # frozen_string_literal: true
 
+require 'optparse'
 require './command'
 
 opt = OptionParser.new
